requirements.txt
fastapi==0.116.1
uvicorn==0.25.0
sqlalchemy==2.0.36
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0
pydantic==2.11.7
pydantic-settings==2.10.1
PyPDF2==3.0.1
python-docx==0.8.11
chromadb==0.4.24
openai==1.51.0
sentence-transformers==2.2.2
numpy==1.24.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4